+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Renamed `plantuml.legend` to `c4plantuml.legend`.
 - Renamed `plantuml.sequenceDiagrams` to `plantuml.sequenceDiagram`.
 - Adds a view/view set property named `c4plantuml.stereotypes` that can be used to enable/disable stereotypes (these are always on by default when the legend is not shown; #29).
+- Adds a `c4plantuml.stdlib` view/view set property to allow users to choose which version of C4-PlantUML should be used (built-in standard library, or GitHub).
 
 ## 1.7.0 (3rd October 2022)